refactor(CouponList.tsx): CouponLIst 컴포넌트를 CouponModal 컴포넌트에서 분리
